+++ src/easysam/prismarine.py
@@ -1,4 +1,5 @@
 from pathlib import Path
+import logging as lg
 
 import prismarine.prisma_client as client
 
@@ -12,6 +13,17 @@ def generate(directory, resources):
     prisma_base = prisma.get('default-base')
     prisma_tables = prisma['tables'] or []
     access_module = prisma.get('access-module') or 'prismarine.runtime.dynamo_default'
+    extra_imports_def = prisma.get('extra-imports')
+
+    if extra_imports_def:
+        extra_imports = [i.split(':') for i in extra_imports_def]
+
+        lg.info('Prismarine imports collected')
+        for i in extra_imports:
+            lg.info(f'Class {i[0]}:{i[1]}')
+    else:
+        extra_imports = []
+        lg.info('No extra prismarine imports')
 
     for prisma_integration in prisma_tables:
         base = prisma_integration.get('base') or prisma_base
@@ -32,7 +44,8 @@ def generate(directory, resources):
             )
 
         content = client.build_client(
-            cluster, base_dir, base, access_module
+            cluster, base_dir, base, access_module,
+            extra_imports=extra_imports
         )
 
         client.write_client(content, base_dir, package)